backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
celery>=5.3.0
redis>=5.0.0
django-cors-headers>=4.3.0
django-celery-beat>=2.5.0
Pillow>=10.0.0

